lizaalert-backend/requirements.txt
asgiref==3.5.0; python_version >= "3.7"
certifi==2021.10.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
cffi==1.15.0; python_version >= "3.7"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3.6"
coreapi==2.3.3; python_version >= "3.6"
coreschema==0.0.4; python_version >= "3.6"
cryptography==36.0.2; python_version >= "3.7"
cssselect2==0.5.0; python_version >= "3.7"
defusedxml==0.7.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
dj-rest-auth==2.2.4; python_version >= "3.5"
django-allauth==0.50.0
django-phonenumber-field==6.1.0; python_version >= "3.7"
django==3.2.12; python_version >= "3.6"
djangorestframework-simplejwt==5.1.0; python_version >= "3.7"
djangorestframework==3.13.1; python_version >= "3.6"
drf-yasg==1.20.0; python_version >= "3.6"
easy-thumbnails==2.8.1; python_version >= "3.6"
environs==9.5.0; python_version >= "3.6"
gunicorn==20.1.0; python_version >= "3.5"
idna==3.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
inflection==0.5.1; python_version >= "3.6"
itypes==1.2.0; python_version >= "3.6"
jinja2==3.1.1; python_version >= "3.7"
lxml==4.8.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
markupsafe==2.1.1; python_version >= "3.7"
marshmallow==3.15.0; python_version >= "3.7"
oauthlib==3.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
packaging==21.3; python_version >= "3.7"
phonenumbers==8.12.45
pillow==9.0.1; python_version >= "3.7" and python_version < "4"
psycopg2-binary==2.9.3; python_version >= "3.6"
pycparser==2.21; python_version >= "3.7" and python_full_version < "3.0.0" or python_version >= "3.7" and python_full_version >= "3.4.0"
pydantic==1.9.0; python_full_version >= "3.6.1"
pyjwt==2.3.0; python_version >= "3.7"
pyparsing==3.0.7; python_version >= "3.7"
python-dotenv==0.20.0; python_version >= "3.6"
python3-openid==3.2.0
pytz==2022.1; python_version >= "3.7"
reportlab==3.6.9; python_version >= "3.7" and python_version < "4"
requests-oauthlib==1.3.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
requests==2.27.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
ruamel.yaml.clib==0.2.6; platform_python_implementation == "CPython" and python_version < "3.11" and python_version >= "3.6"
ruamel.yaml==0.17.21; python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.7"
svglib==1.2.1; python_version >= "3.7"
tinycss2==1.1.1; python_version >= "3.7"
typing-extensions==4.1.1; python_version >= "3.7" and python_full_version >= "3.6.1" and python_version < "3.8"
uritemplate==4.1.1; python_version >= "3.6"
urllib3==1.26.9; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
webencodings==0.5.1; python_version >= "3.7"
